// frontend/src/api/generated/types.gen.ts
export type ProfileWrite = {
    experienceLevel: 'beginner' | 'intermediate' | 'advanced' | 'expert';
    displayUnits: 'metric' | 'imperial';
};

export type ProfileUpdate = {
    experienceLevel: 'beginner' | 'intermediate' | 'advanced' | 'expert';
    displayUnits: 'metric' | 'imperial';
    expectedVersion: number;
};

export type Profile = {
    experienceLevel: 'beginner' | 'intermediate' | 'advanced' | 'expert';
    displayUnits: 'metric' | 'imperial';
    version: number;
    createdAt: string;
    updatedAt: string;
};